+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-- Fixed [Get-OGP][] Stop scraping amazon.co.jp.
+- Changed [Group-Aggregate][] Separate the C# file.
+- Fixed [Get-OGP][] Disallow scraping of amazon.co.jp.
 - Fixed [Get-LabelBlock][] skip trim when comment field is empty.
 - Added [Get-LabelBlock][] function.
 - Added [glip][] function.